Refactor: 파일명 수정 및 코드 리팩토링
