--- src/components/titleBorder.jsx
@@ -1,30 +1,0 @@
-import styled from "styled-components";
-
-
-const StyledTitle = styled.div`
-  font-size:2rem;
-  font-weight:bolder;
-  color: #3d3d3d;
- 
-
-  text-shadow: calc(3px * 1) 0 0 white,
-    calc(3px * 0.9239) calc(3px * 0.3827) 0 white,
-    calc(3px * 0.7071) calc(3px * 0.7071) 0 white,
-    calc(3px * 0.3827) calc(3px * 0.9239) 0 white,
-    0 calc(3px * 1) 0 white,
-    calc(3px * -0.3827) calc(3px * 0.9239) 0 white,
-    calc(3px * -0.7071) calc(3px * 0.7071) 0 white,
-    calc(3px * -0.9239) calc(3px * 0.3827) 0 white,
-    calc(3px * -1) 0 0 white,
-    calc(3px * -0.9239) calc(3px * -0.3827) 0 white,
-    calc(3px * -0.7071) calc(3px * -0.7071) 0 white,
-    calc(3px * -0.3827) calc(3px * -0.9239) 0 white,
-    0 calc(3px * -1) 0 white,
-    calc(3px * 0.3827) calc(3px * -0.9239) 0 white,
-    calc(3px * 0.7071) calc(3px * -0.7071) 0 white,
-    calc(3px * 0.9239) calc(3px * -0.3827) 0 white
-`;
-
-export default function TitleBorder({ children }) {
-    return <StyledTitle>{children}</StyledTitle>;
-}
