+++ src/config/ApiConfig.jsx
@@ -1,0 +1,1 @@
+export const SERVER_URL = "http://3.37.89.60:8080";
