+++ src/index.js
@@ -2,8 +2,8 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import "../src/pages/index.css";
-import Login from "./pages/member/login";
-import SignUp from "./pages/member/signUp";
+import LoginPage from "./pages/member/LoginPage";
+import SignUpPage from "./pages/member/SignUpPage";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
@@ -12,9 +12,9 @@ root.render(
 
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<Login/>} />
-        <Route path="/login" element={<Login/>} />
-        <Route path="/signUp" element={<SignUp/>} />
+        <Route path="/" element={<LoginPage/>} />
+        <Route path="/login" element={<LoginPage/>} />
+        <Route path="/signup" element={<SignUpPage/>} />
       </Routes>
     </BrowserRouter>
   </React.StrictMode>
